requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
requests
python-dotenv
flask-cors
supabase
postgrest
realtime
storage3
gotrue
httpx
